Move everything over from `middle::const_val` to `mir::interpret`
